+++ citadel/logging.c
@@ -28,6 +28,6 @@ void rec_log(unsigned int lrtype, char *name) {
 
 	time(&now);
 	fp = fopen("citadel.log", "a");
-	fprintf(fp, "%ld|%u|%s\n", now, lrtype, name);
+	fprintf(fp, "%ld|%u|%s\n", (long)now, lrtype, name);
 	fclose(fp);
 	}
